requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
PyYAML==6.0
requests==2.26.0
tomark==0.1.4
urllib3==1.26.7
